+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,6 +1,6 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.dto.GroupDTO;
+import com.kodilla.ecommercee.domain.dto.GroupDTO;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
