+++ src/main/java/com/kodilla/ecommercee/controller/ProductController.java
@@ -1,6 +1,6 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.dto.ProductDTO;
+import com.kodilla.ecommercee.domain.dto.ProductDTO;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
