+++ src/main/java/com/kodilla/ecommercee/domain/dto/GroupDTO.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.dto;
+package com.kodilla.ecommercee.domain.dto;
 
 public class GroupDTO {
