+++ src/main/java/com/kodilla/ecommercee/domain/dto/UserKeyDTO.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.domain.dto;
+
+public class UserKeyDTO {
+}
