--- src/main/java/com/kodilla/ecommercee/dto/ProductDTO.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.dto;
-
-public class ProductDTO {
-}
